Feature: If some components somehow manage to get "off the grid," they are relocated back onto the grid once they are moved.
Bug fix: Several types of exceptions could occur during simulation, particularly when poking the simulation while the simulation is busy. Many of these have been removed.

Bug fix: When editing the appearance and copying to the clipboard, the anchor's location and facing are stored to the clipboard if the selection includes the anchor.


git-svn-id: svn://svn.code.sf.net/p/circuit/code/trunk@244 70edf91d-0b9e-4248-82e7-2488d7716404